+++ pajbot/modules/chat_alerts/newchatteralert.py
@@ -1,8 +1,16 @@
+from __future__ import annotations
+
+from typing import TYPE_CHECKING, Dict, Optional
+
 import logging
 
 from pajbot.managers.handler import HandlerManager
+from pajbot.models.user import User
 from pajbot.modules import BaseModule, ModuleSetting
 from pajbot.modules.chat_alerts import ChatAlertModule
+
+if TYPE_CHECKING:
+    from pajbot.bot import Bot
 
 log = logging.getLogger(__name__)
 
@@ -75,14 +83,16 @@ class NewChatterAlertModule(BaseModule):
         ),
     ]
 
-    def __init__(self, bot):
+    def __init__(self, bot: Optional[Bot]) -> None:
         super().__init__(bot)
 
-    def on_new_chatter(self, user):
+    def on_new_chatter(self, user: User) -> None:
         """
         A user just announced they are a new chatter.
         Send the event to the websocket manager and send a customized message in chat/whispers.
         """
+
+        assert self.bot is not None
 
         payload = {"username": user.name}
         self.bot.websocket_manager.emit("new_chatter", payload)
@@ -106,18 +116,22 @@ class NewChatterAlertModule(BaseModule):
         if alert_message != "":
             self.bot.say(alert_message.format(user=user, points=self.settings["grant_points_on_new_chatter"]))
 
-    def on_usernotice(self, source, tags, **rest):
+    def on_usernotice(self, source: User, message: str, tags: Dict[str, str]) -> bool:
         if "msg-id" not in tags:
-            return
+            return True
 
         if tags["msg-id"] == "ritual":
             if "display-name" not in tags:
                 log.debug(f"newchatteralert requires a display-name, but it is missing: {tags}")
-                return
+                return True
             self.on_new_chatter(source)
 
-    def enable(self, bot):
-        HandlerManager.add_handler("on_usernotice", self.on_usernotice)
+        return True
 
-    def disable(self, bot):
-        HandlerManager.remove_handler("on_usernotice", self.on_usernotice)
+    def enable(self, bot: Optional[Bot]) -> None:
+        if bot is not None:
+            HandlerManager.add_handler("on_usernotice", self.on_usernotice)
+
+    def disable(self, bot: Optional[Bot]) -> None:
+        if bot is not None:
+            HandlerManager.remove_handler("on_usernotice", self.on_usernotice)